♻️ PostImages 컴포넌트 삭제 후 PostAlbum으로 통합 🐛 이미지 선택 삭제 시 버그 수정, 30장 제한 기능 보완
